refactored resend modal and button, decline modal and button into new components
