+++ packages/app/src/components/nominationBanner/declineAppBtn.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+const DeclineAppBtn = (props) => {
+
+  return ( 
+      <button
+        disabled={props.status == 'Declined'}
+        className="decline-button banner-buttons"
+        onClick={props.toggleDeclineAppModalState}
+      >
+        Decline Application
+      </button>
+    )
+}
+
+
+export default DeclineAppBtn;
